+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-## [1.0.2] - 15.10.2019.
+## 1.0.3 - 16.10.2019
+* Removed generic event parameter from onTriggered function, its not needed
+
+## 1.0.2 - 15.10.2019
 
 * Updated dependencies
 * Provide example
 
-## [1.0.1] - 13.10.2019.
+## 1.0.1 - 13.10.2019
 
 * Changes for pub.dev
 
-## [1.0.0] - 13.10.2019.
+## 1.0.0 - 13.10.2019
 
 * Added initial functionality